// apps/web/src/api/model/apiHydratedSecretKey.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { ApiHydratedSecretKeyApiVersion } from './apiHydratedSecretKeyApiVersion';
import type { ApiHydratedSecretKeyScopesItem } from './apiHydratedSecretKeyScopesItem';
import type { ApiHydratedSecretKeyStatus } from './apiHydratedSecretKeyStatus';

export interface ApiHydratedSecretKey {
  /** Secret key name */
  name: string;
  /** Generated secret key value */
  key: string;
  /** Allowed scopes for this key */
  scopes: ApiHydratedSecretKeyScopesItem[];
  /** API version assigned to this key */
  apiVersion?: ApiHydratedSecretKeyApiVersion;
  /** Current status of the key */
  status?: ApiHydratedSecretKeyStatus;
  /** Whether GRP is enabled for this key */
  grpEnabled?: boolean;
  /** CIDR whitelist for this key */
  cidrWhitelist?: string[];
  /** Business ID */
  business: string;
  /** Last time the key was used */
  lastUsed?: string;
  /** Document id */
  id: string;
  /** Document created at */
  createdAt: string;
  /** Document updated at */
  updatedAt: string;
}