// apps/web/src/api/model/secretKeyMetricsResponseDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { SecretKeyDailyStats } from './secretKeyDailyStats';

export interface SecretKeyMetricsResponseDto {
  /** Requests made today */
  requestsToday: number;
  /** Errors today (responses with status code >= 400) */
  errorsToday: number;
  /** Total lifetime requests */
  totalRequests: number;
  /** Daily request breakdown for requested period */
  dailyStats: SecretKeyDailyStats[];
}